requirements.txt
flask>=3.0
pymongo>=4.6
python-nmap>=0.7
dnspython>=2.5
requests>=2.31
